[WIP] Refactoring for typescript
