+++ tslint.json
@@ -1,0 +1,81 @@
+{
+	"rules": {
+		"ban": [true,
+			["angular", "forEach"]
+		],
+		"class-name": true,
+		"comment-format": [false,
+			"check-space",
+			"check-lowercase"
+		],
+		"curly": true,
+		"eofline": false,
+		"forin": false,
+		"indent": [true, 4],
+		"interface-name": [true, "never-prefix"],
+		"jsdoc-format": true,
+		"label-position": true,
+		"label-undefined": true,
+		"max-line-length": [false, 140],
+		"no-any": false,
+		"no-arg": true,
+		"no-bitwise": true,
+		"no-console": [false,
+			"debug",
+			"info",
+			"time",
+			"timeEnd",
+			"trace"
+		],
+		"no-consecutive-blank-lines": false,
+		"no-construct": true,
+		"no-debugger": false,
+		"no-duplicate-key": true,
+		"no-duplicate-variable": true,
+		"no-empty": false,
+		"no-eval": true,
+		"no-string-literal": false,
+		"trailing-comma": true,
+		"no-trailing-whitespace": true,
+		"no-unreachable": true,
+		"no-unused-expression": false,
+		"no-unused-variable": true,
+		"no-use-before-declare": true,
+		"no-var-requires": false,
+		"one-line": [true,
+			"check-catch",
+			"check-else",
+			"check-open-brace",
+			"check-whitespace"
+		],
+		"quotemark": [false, "double"],
+		"radix": false,
+		"semicolon": true,
+		"triple-equals": [true, "allow-null-check"],
+		"typedef": [false,
+			"callSignature",
+			"catchClause",
+			"indexSignature",
+			"parameter",
+			"propertySignature",
+			"variableDeclarator"
+		],
+		"typedef-whitespace": [false,
+			["callSignature", "noSpace"],
+			["catchClause", "noSpace"],
+			["indexSignature", "space"]
+		],
+		"use-strict": [true,
+			"check-module",
+			"check-function"
+		],
+		"variable-name": [true, "allow-leading-underscore"],
+		"whitespace": [false,
+			"check-branch",
+			"check-decl",
+			"check-operator",
+			"check-separator",
+			"check-type"
+		]
+	}
+}
